[IMP] [REF] website: introducing a mixin class to handle website_published and website_url fields. This will allow to remove custom code in every website addons that redefine those two fields and the URL getter method.
The published button in backend has also been cleaned. It is now a link button
that redirects to the website_url. It does not change the publish status
anymore. Indeed it makes sense publish and unpublish website documents
directly in the website, while checking that everything is correct, instead
of doing it in the backend.

Next commit will update the Odoo addons to use this mixin and the new button
in the form view.