[IMP] stock: adapt views of current inventory valuation and stock level forecast
The stock level forecast view should have inverse columns and rows.  The line graph
is also better by default.

The stock valuation view should only be in reporting (removed the graph part) and is replaced by a pivot and
graph view of the quants in the warehouse menu.  We only group by location by default.