[IMP] website: only load website.xml when needed
Rewriting of the ill-fated e18a3a4 after c6d552e smashed it to the
ground with very good reasons.

I moved all the templates that were only used in edit mode to the
website.editor.xml file and the website.prompt function in the
editor as it was only used by editor javascript files anyway.

Now website.xml only contains website.error_dialog, and is lazily
loaded the first time an error needs to be displayed to the user.